requirements.txt

cssselect2==0.7.0
Django==5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
html5lib==1.1
idna==3.8
lxml==5.3.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.1
pypdf==4.3.1
pypng==0.20220715.0
pytest==8.3.2
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
xhtml2pdf==0.2.16
